refactor(map): decompose the 5,678-line MapViewController god file
Mechanical extraction into sibling files in Features/Map/Controllers —
no behavior change:

- LocationManager.swift — the shared CLLocation source
- TacticalMapView.swift — 2D Mapbox representable + coordinator + the
  radial-menu Mapbox bridge
- CesiumMainMap.swift — 3D Cesium WKWebView representable + JS bridge
- MapDrawingAnnotations.swift — MKAnnotation classes for drawing shapes
- ATAKMapChrome.swift — ATAKStatusBar / ATAKBottomToolbar / ATAKSidePanel /
  OverlaySettingsPanel + button helpers

The embedded ~300-line Cesium scene HTML moves out of the Swift string
literal into a bundled resource (OmniTAKMobile/Resources/cesium_scene.html,
same __CESIUM_ION_TOKEN__ placeholder convention as the Android asset);
CesiumMainMap.html now loads it at runtime and injects the Ion token from
CesiumIonConfig (Config.xcconfig → Info.plist), preserving the
never-hardcode-the-token invariant from the security pass.

ATAKMapView and its body stay in MapViewController.swift (now 2,208
lines). This was the single largest maintainability/bus-factor item in
the acquisition audit: one file held both map engines, the GPS source,
the chrome, and the scene HTML.
